python/plant-disease-detection-api/requirements.txt
numpy==1.17.2
Pillow==6.1.0
six==1.12.0
torch==1.2.0+cpu
torchvision==0.4.0+cpu
